Encapsulate the blacklight config
All values set on Blacklight::Engine.config are actually stored
in a class variable on Railtie::Configuration. So these pollute
that namespace.  All other railties are encapsulating their configs
onto a single entry in that scope.  This change makes Blacklight do
likewise.

This probably needs to be a major revision